Address CirclePacking PR review comments
- Reuse the generic NodesPlot/LabelsPlot components for CirclePacking's
  node/label rendering, the same way DendrogramPlot already does -
  positions/layout (pack vs cluster) were the only real difference.
  CirclePackingBase is gone; a new useCirclePackingLayout hook computes
  the pack layout once, shared by both plots, each now rendering to its
  own Canvas/SVG layer.
- Rename the stale "should render without needing an external
  RectangularChart wrapper" unit test (RectangularChart was removed
  before this PR) to "should render correctly".
